requirements.txt
selenium
webdriver-manager
python-dotenv
cryptography
pickle-mixin
